// src/store/main/system/types.ts
/*
 * @Author: Cyan_Breeze
 * @Description:store中system模块中state的type
 * @Date: 2022-10-13 22:35:49
 * @LastEditTime: 2022-11-14 16:15:17
 * @FilePath: \vue3-cms\src\store\main\system\types.ts
 */
export interface ISystemState {
  usersList: any[]
  usersCount: number
  roleList: any[]
  roleCount: number
  goodsList: any[]
  goodsCount: number
  menuList: any[]
  menuCount: number
}